Replace file output parser by recent value socket
pylarexx can write the sensor data to a file
or give the last received value as ouput on
a network socket.
This patch uses the socket to read the sensor
data from and makes the file writing obsolete.
Especially on raspberry pi with sd card this can
be usefull to reduce wear of flash memory.