Bugfix: sensor name must not start with digit
The former release used the ID as given by the arexx sensors.
In Munin the sensor name must not start with a digit, which
lead to strange problems (picture created in future and thus not
updated, rrd file error etc).
With this update sensor name is preceded by "ax" for arexx.

Warning: as new rrd-files will be created former data is not
visible any longer.